requirements.txt
numpy==1.18.5
pandas
matplotlib
scipy==1.4.1
keras==2.3.1
PyYaml
RootInteractive
tf-nightly-gpu==2.4.0.dev20200908; sys_platform == 'linux'
pydot
pylint
